change sidebar to topbar
