+++ frontend/zeevision/src/pages/JobsPage.tsx
